client/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
PyJWT>=2.8.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
psutil>=5.9.0
pydantic>=2.0.0
